refactor make_skippable tests by introducing generic_make_skippable_tests
