+++ tests/test_make_skippable.cpp
@@ -9,218 +9,80 @@
 #include <gtest/gtest.h>
 
 #include "funkypipes/details/make_skippable.hpp"
+#include "predefined/execution_semantics/make_skippable_tests.hpp"
 
 using namespace funkypipes;
 using namespace funkypipes::details;
+using namespace funkypipes::test;
 
-TEST(MakeSkippable, callable_having_value_argument__called_with_lvalue_optional__works) {
-  auto lambda = [](int value) { return std::to_string(value); };
+auto makeSkippableFn = [](auto&&... args) { return makeSkippable(std::forward<decltype(args)>(args)...); };
 
-  auto skippable_lambda = makeSkippable(lambda);
-
-  std::optional<int> argument{1};
-  std::optional<std::string> res = skippable_lambda(argument);
-  EXPECT_TRUE(res.has_value());
-  EXPECT_EQ(res.value(), "1");
+// feature: callables - generic callables
+TEST(MakeSkippable, genericFunctorCallable_calledWithValue_isExecuted) {
+  ASSERT_NO_FATAL_FAILURE(execution_semantics::genericFunctorCallable_calledWithValue_isExecuted(makeSkippableFn));
 }
 
-TEST(MakeSkippable, callable_having_value_argument__called_with_rvalue_optional__works) {
-  auto lambda = [](int value) { return std::to_string(value); };
-
-  auto skippable_lambda = makeSkippable(lambda);
-
-  std::optional<int> argument{1};
-  std::optional<std::string> res = skippable_lambda(std::move(argument));
-  EXPECT_TRUE(res.has_value());
-  EXPECT_EQ(res.value(), "1");
+TEST(MakeSkippable, genericLambdaCallable_calledWithValue_isExecuted) {
+  ASSERT_NO_FATAL_FAILURE(execution_semantics::genericLambdaCallable_calledWithValue_isExecuted(makeSkippableFn));
 }
 
-TEST(MakeSkippable, callable_having_value_argument__called_with_non_copyable_rvalue_optional__works) {
-  struct NonCopyableArg {
-    ~NonCopyableArg() = default;
-    NonCopyableArg(const NonCopyableArg&) = delete;
-    NonCopyableArg(NonCopyableArg&&) = default;
-    NonCopyableArg& operator=(const NonCopyableArg&) = delete;
-    NonCopyableArg& operator=(NonCopyableArg&&) = delete;
-  };
-
-  auto lambda = [](NonCopyableArg arg) { return arg; };
-
-  auto skippable_lambda = makeSkippable(lambda);
-
-  auto argument = std::make_optional(NonCopyableArg{});
-  std::optional<NonCopyableArg> res = skippable_lambda(std::move(argument));
-  EXPECT_TRUE(res.has_value());
+TEST(MakeSkippable, overloadedFunctorCallable_calledWithValue_isExecuted) {
+  ASSERT_NO_FATAL_FAILURE(execution_semantics::overloadedFunctorCallable_calledWithValue_isExecuted(makeSkippableFn));
 }
 
-TEST(MakeSkippable, callable_having_reference_argument__called_with_rvalue_optional__works) {
-  auto lambda = [](const int& value) { return std::to_string(value); };
-
-  auto skippable_lambda = makeSkippable(lambda);
-
-  std::optional<int> argument{1};
-  std::optional<std::string> res = skippable_lambda(std::move(argument));
-  EXPECT_TRUE(res.has_value());
-  EXPECT_EQ(res.value(), "1");
+// feature: callables - move only
+TEST(MakeSkippable, nonCopyableCallable_calledWithValue_isExecuted) {
+  ASSERT_NO_FATAL_FAILURE(execution_semantics::nonCopyableCallable_calledWithValue_isExecuted(makeSkippableFn));
 }
 
-TEST(MakeSkippable, callable_returning_value__called_with_nullopt__is_skipped) {
-  auto lambda = [](int value) {
-    throw std::exception{};
-    return std::to_string(value);
-  };
-  auto skippable_lambda = makeSkippable(lambda);
-
-  std::optional<int> argument{std::nullopt};
-  std::optional<std::string> res;
-  ASSERT_NO_THROW(res = skippable_lambda(argument));
-  EXPECT_FALSE(res.has_value());
+// feature: data - value categories
+TEST(MakeSkippable, callableHavingValueArgument_calledWithLValueOptional_works) {
+  ASSERT_NO_FATAL_FAILURE(
+      execution_semantics::callableHavingValueArgument_calledWithLValueOptional_works(makeSkippableFn));
 }
 
-TEST(MakeSkippable, callable_returning_optional__called_with_nullopt__is_skipped) {
-  auto lambda = [](int value) {
-    throw std::exception{};
-    return std::make_optional(std::to_string(value));
-  };
-  auto skippable_lambda = makeSkippable(lambda);
-
-  std::optional<int> argument{std::nullopt};
-  std::optional<std::string> res;
-  ASSERT_NO_THROW(res = skippable_lambda(argument));
-  EXPECT_FALSE(res.has_value());
+TEST(MakeSkippable, callableHavingValueArgument_calledWithRValueOptional_works) {
+  ASSERT_NO_FATAL_FAILURE(
+      execution_semantics::callableHavingValueArgument_calledWithRValueOptional_works(makeSkippableFn));
 }
 
-TEST(MakeSkippable, callable_returning_optional__called_with_value__is_executed) {
-  auto lambda = [](int value) { return std::make_optional(std::to_string(value)); };
-
-  auto skippable_lambda = makeSkippable(lambda);
-
-  std::optional<int> argument{1};
-  std::optional<std::string> res = skippable_lambda(argument);
-  EXPECT_TRUE(res.has_value());
-  EXPECT_EQ(res.value(), "1");
+TEST(MakeSkippable, callableHavingReferenceArgument_calledWithRValueOptional_works) {
+  ASSERT_NO_FATAL_FAILURE(
+      execution_semantics::callableHavingReferenceArgument_calledWithRValueOptional_works(makeSkippableFn));
 }
 
-struct GenericFn {
-  template <typename TArg, typename = std::enable_if_t<not IsOptional<TArg>::value>>
-  auto operator()(TArg arg) {
-    return arg;
-  }
-};
-
-TEST(MakeSkippable, generic_functor_callable__called_with_value__is_executed) {
-  auto skippable_generic_functor = makeSkippable(GenericFn{});
-
-  {
-    std::optional<int> argument{1};
-    std::optional<int> res = skippable_generic_functor(argument);
-    EXPECT_TRUE(res.has_value());
-    EXPECT_EQ(res.value(), 1);
-  }
-  {
-    auto argument = std::make_optional("string");
-    std::optional<std::string> res = skippable_generic_functor(argument);
-    EXPECT_TRUE(res.has_value());
-    EXPECT_EQ(res.value(), "string");
-  }
+TEST(MakeSkippable, callableProvidingConstReference_called_returnsOptionalConstReference) {
+  ASSERT_NO_FATAL_FAILURE(
+      execution_semantics::callableProvidingConstReference_called_returnsOptionalConstReference(makeSkippableFn));
 }
 
-TEST(MakeSkippable, generic_lambda_callable__called_with_value__is_executed) {
-  auto lambda = [](auto arg) { return arg; };
-
-  auto skippable_lambda = makeSkippable(lambda);
-
-  {
-    std::optional<int> argument{1};
-    std::optional<int> res = skippable_lambda(argument);
-    EXPECT_TRUE(res.has_value());
-    EXPECT_EQ(res.value(), 1);
-  }
-  {
-    auto argument = std::make_optional("string");
-    std::optional<std::string> res = skippable_lambda(argument);
-    EXPECT_TRUE(res.has_value());
-    EXPECT_EQ(res.value(), "string");
-  }
+TEST(MakeSkippable, callableProvidingReference_called_returnsOptionalReference) {
+  ASSERT_NO_FATAL_FAILURE(
+      execution_semantics::callableProvidingReference_called_returnsOptionalReference(makeSkippableFn));
 }
 
-TEST(MakeSkippable, overloaded_functor_callable__called_with_value__is_executed) {
-  struct OverloadedFn {
-    int operator()(int input) const { return input; }
-    std::string operator()(std::string input) const { return input; }
-  };
-
-  auto skippable_overloaded_fn = makeSkippable(OverloadedFn{});
-
-  {
-    std::optional<int> argument{1};
-    std::optional<int> res = skippable_overloaded_fn(argument);
-    EXPECT_TRUE(res.has_value());
-    EXPECT_EQ(res.value(), 1);
-  }
-
-  {
-    auto argument = std::make_optional("1");
-    std::optional<std::string> res = skippable_overloaded_fn(argument);
-    EXPECT_TRUE(res.has_value());
-    EXPECT_EQ(res.value(), "1");
-  }
+// feature: data - move only
+TEST(MakeSkippable, callableHavingValueArgument_calledWithNonCopyableRValueOptional_works) {
+  ASSERT_NO_FATAL_FAILURE(
+      execution_semantics::callableHavingValueArgument_calledWithNonCopyableRValueOptional_works(makeSkippableFn));
 }
 
-TEST(MakeSkippable, non_copyable_callable__called_with_value__is_executed) {
-  struct NonCopyableFn {
-    ~NonCopyableFn() = default;
-    NonCopyableFn(const NonCopyableFn&) = delete;
-    NonCopyableFn(NonCopyableFn&&) = default;
-    NonCopyableFn& operator=(const NonCopyableFn&) = delete;
-    NonCopyableFn& operator=(NonCopyableFn&&) = delete;
-
-    int operator()(int value) const { return value; }
-  };
-
-  auto skippable_fn = makeSkippable(NonCopyableFn{});
-
-  std::optional<int> argument{1};
-  std::optional<int> res = skippable_fn(argument);
-  EXPECT_TRUE(res.has_value());
-  EXPECT_EQ(res.value(), 1);
+// feature: std::optional support
+TEST(MakeSkippable, callableReturningOptional_calledWithValue_isExecuted) {
+  ASSERT_NO_FATAL_FAILURE(execution_semantics::callableReturningOptional_calledWithValue_isExecuted(makeSkippableFn));
 }
 
-TEST(MakeSkippable, callable_providing_const_reference__called__returns_optional_const_reference) {
-  auto lambda = [](const int& arg) -> const int& { return arg; };
-  auto skippable_lambda = makeSkippable(lambda);
-
-  int argument_raw = 1;
-  std::optional<std::reference_wrapper<const int>> argument{argument_raw};
-  std::optional<std::reference_wrapper<const int>> res = skippable_lambda(argument);
-  EXPECT_TRUE(res.has_value());
-  EXPECT_EQ(res.value(), 1);
-
-  const int& result_raw = res.value();
-  EXPECT_EQ(result_raw, 1);
-
-  argument_raw++;
-  EXPECT_EQ(result_raw, 2);
+// feature: chain breaking
+TEST(MakeSkippable, callableReturningValue_calledWithNullopt_isSkipped) {
+  ASSERT_NO_FATAL_FAILURE(execution_semantics::callableReturningValue_calledWithNullopt_isSkipped(makeSkippableFn));
 }
 
-TEST(MakeSkippable, callable_providing_reference__called__returns_optional_reference) {
-  auto lambda = [](int& arg) -> int& { return arg; };
-  auto skippable_lambda = makeSkippable(lambda);
-
-  int argument_raw = 1;
-  std::optional<std::reference_wrapper<int>> argument{argument_raw};
-  std::optional<std::reference_wrapper<int>> res = skippable_lambda(argument);
-  EXPECT_TRUE(res.has_value());
-
-  int& result_raw = res.value();
-  EXPECT_EQ(result_raw, 1);
-
-  result_raw++;
-  EXPECT_EQ(argument_raw, 2);
+TEST(MakeSkippable, callableReturningOptional_calledWithNullopt_isSkipped) {
+  ASSERT_NO_FATAL_FAILURE(execution_semantics::callableReturningOptional_calledWithNullopt_isSkipped(makeSkippableFn));
 }
 
 // TEST(MakeSkippable,
-// callable_called_with_unsupported_argument__triggers_static_assert) {
+// callableCalledWithUnsupportedArgument_triggersStaticAssert) {
 //   auto lambda = [](int value) { return value; };
 //
 //   auto skippable_lambda = makeSkippable(lambda);
